Create node type (#33)
* Implement createMathNode and createConstNode

* Create NodeType

* Please linter and change createFunc signature

createFunc is now (node: Node) => Node because old code would generate
circular dependency with Engine.

* Move NodeType to a separate file

So that we can use the type in engine.

* Dynamically bind StockNodeTypeDeclarations

Provide a function to do so. It can't be done in the engine because of
circular dependencies.

* Create StandardEngine

* Privatize nodeTypeDefinitions and move tests

Since the old engine tests only check for types, it can be renamed
as a StandardEngine test.

* Fix unknown type in math node

* Add stock value types

Also implements barebones addValueTypeDefinition to make this possible.

* Polish add(Node/Value)TypeDefinition documentation

* Remove obsolete createEngine

* Rename ..TypeDeclaration to ..TypeDefinition

Initially it was "definition" but for some reason I wrote them
"declaration" at some places and I believe the naming has to be consistent.

* Swap Engine with StandardEngine in general test

Because it also tests for standard types which is included in
standardEngine but not in plain Engine.

* Handle duplicates in addValueTypeDefinition

* Handle duplicates in addNodeTypeDefinition

* Implement getNodeTypeDefinition

* Implement halt

* Test halting

* Fix linter issues

* Rename Stock types to Standard types

* Get engine implicitly in node

Instead of passing engine explicitly, it's taken from network.