"nothing" and "isSet" changes
From now on no value object implies nothing state which shows the mutual
exclusion between having a value and having "nothing".

hasValue now is renamed to "isSetVariable" to indicate the similarity
with "isSet" and to reduce ambiguity of "hasValue" (which is not an
exactly correct term)

Missing visibility modifiers have been added for explicitness.