Make sockets demand Value constructors
This new change makes it possible to declare the type without passing a
value. Also users don't have to wrap a value with Value<T> which makes
the sockets easy to use. This is made even easier in Node where addInput
and addOutput ask for name which then is used to retrieve a
corresponding Value constructor (or signal an error).

Some changes made in nodeTypeDefinitions to reflect the new API. It's
not transformed fully because there is no standard Value type yet.