Valuetype refactor (#39)
* Turn ValueTypes into classes

* Rename toJson to toJSON

* Fix breaking changes from modifying ValueType

* Resolve errors and tests

* Create Value class and standard values

* Add "nothing" to Value

* Make Value wrap any* value not JsonValue

* Not really any, the ones that can be converted to JSON.

Also includes improved toJSON().

* Adjust value checking process

* Fix standard value definitions

* Add value set/get and setNothing()

* Remove "nothing" from Value

* "nothing" and "isSet" changes

From now on no value object implies nothing state which shows the mutual
exclusion between having a value and having "nothing".

hasValue now is renamed to "isSetVariable" to indicate the similarity
with "isSet" and to reduce ambiguity of "hasValue" (which is not an
exactly correct term)

Missing visibility modifiers have been added for explicitness.

* Add setNothing()

* Partially update inputSocket

In a similar way Socket was changed recently

* Fix Outputsocket like Socket

* Fix remaining socket issues

This includes fixing the return type of getters.

* Fix some issues caused by missing generic types

unknown is passed as generic because a socketCollection can consist of
sockets with different types

* Add some detail to isSetVariable documentation

This includes the importand distinction of value = null (aka nothing)
and isSetVariable = false

* Implement valueDefinitions

* Create standard value definitions

* Move ValueConstructor to value.ts

* Fix some inconsistencies

* Rename unset(reset) to clear and init to reset

"unset" was "reset" before but it should have been called "clear". This
includes not only changes in sockets but also node and network which
followed a similar naming convention.

* Swap unknown assignments to Value<unknown>

* Make sockets demand Value constructors

This new change makes it possible to declare the type without passing a
value. Also users don't have to wrap a value with Value<T> which makes
the sockets easy to use. This is made even easier in Node where addInput
and addOutput ask for name which then is used to retrieve a
corresponding Value constructor (or signal an error).

Some changes made in nodeTypeDefinitions to reflect the new API. It's
not transformed fully because there is no standard Value type yet.

* Add default Value type

* Fix tests and add tiny features

Features include setDefaultNothing for inputSocket and isWaiting for
outputSocket.

Tests also include examples. The quadraticForumalaNetworkTest was
suspended because it's bloated.

* Fix node reset() recursion

* Delegate default value type business to socket

This is done so that socket can be used without an engine.
commit WIP

* Make node ignore engine if no type name provided

Before this commit Node always checked for engine but for default
type (when name is omitted) the engine is not necessary.

A helper has been created to avoid duplicating code in addInput and addOutput.

* Fix linter issues

* Remove ValueType

* Fix bug where default value should have been taken

* Rename ValueTypeExtractor to getValueType

* Rename Value(Constructor) to SocketValue(Type)

* Remove value from socket constructor + minor fixes

This change also affects Node's addInput and addOutput that mirror
socket constructor.

Minor fixes include changing Value(Constructor) to SocketValue(Type).

* Change old .getInput().value to getInput()

Some of them are not fixed because they actually need access to that SocketValue object

* Implement getSocketValue

* Fix minor linter and naming issues

* Test various socket values and fix isJSON()

* Test and tweak NumberSocketValue

This also includes some modifications in SocketValue that calls static
methods dynamically (it uses this.constructor to find the classname).

Right now NaN and +-Infinity is accepted at number but not in normal SocketValue.

* Fix broken tests by fixing e.target.getValue()

The problem is that sockets were unexpectedly passed undefined, even
though they had actual values. After skimming through the stack trace I
noticed getValue().value which shouldn't be correct and as I discovered
e.target is of type any.

* Create a folder for socket tests

* Test default socket type assignment

* Fix linter issues

* Pass values as strings as a temporary solution

* Fix the async test